Add support for applications based on the Rivermax Application Library
The Rivermax Application Library can be used to implement
application-level logic that is not covered by the low level Rivermax
API, thereby providing for easier and faster integration with Rivermax.
The Rivermax Application Library is provided alongside the Rivermax SDK.

Added a generic target-library that provides everything required for the
creation of an application which is based on the Rivermax Application
Library.

Signed-off-by: Arseniy Aharonov <aaharonov@nvidia.com>